requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
openai==2.14.0
chromadb==0.4.22
python-dotenv==1.0.0
pandas==2.2.0

